# drivers/iovec: revert vector io implement from loop/null/zero driver