--- a/drivers/misc/dev_zero.c
+++ b/drivers/misc/dev_zero.c
@@ -40,8 +40,10 @@
  * Private Function Prototypes
  ****************************************************************************/
 
-static ssize_t devzero_readv(FAR struct file *filep, FAR struct uio *uio);
-static ssize_t devzero_writev(FAR struct file *filep, FAR struct uio *uio);
+static ssize_t devzero_read(FAR struct file *filep, FAR char *buffer,
+                            size_t buflen);
+static ssize_t devzero_write(FAR struct file *filep, FAR const char *buffer,
+                             size_t buflen);
 static int     devzero_poll(FAR struct file *filep, FAR struct pollfd *fds,
                             bool setup);
 
@@ -51,17 +53,17 @@ static int     devzero_poll(FAR struct file *filep, FAR struct pollfd *fds,
 
 static const struct file_operations g_devzero_fops =
 {
-  NULL,           /* open */
-  NULL,           /* close */
-  NULL,           /* read */
-  NULL,           /* write */
-  NULL,           /* seek */
-  NULL,           /* ioctl */
-  NULL,           /* mmap */
-  NULL,           /* truncate */
-  devzero_poll,   /* poll */
-  devzero_readv,  /* readv */
-  devzero_writev  /* writev */
+  NULL,          /* open */
+  NULL,          /* close */
+  devzero_read,  /* read */
+  devzero_write, /* write */
+  NULL,          /* seek */
+  NULL,          /* ioctl */
+  NULL,          /* mmap */
+  NULL,          /* truncate */
+  devzero_poll,  /* poll */
+  NULL,          /* readv */
+  NULL           /* writev */
 };
 
 /****************************************************************************
@@ -69,41 +71,29 @@ static const struct file_operations g_devzero_fops =
  ****************************************************************************/
 
 /****************************************************************************
- * Name: devzero_readv
+ * Name: devzero_read
  ****************************************************************************/
 
-static ssize_t devzero_readv(FAR struct file *filep, FAR struct uio *uio)
+static ssize_t devzero_read(FAR struct file *filep, FAR char *buffer,
+                            size_t len)
 {
-  size_t total = uio->uio_resid;
-  FAR const struct iovec *iov = uio->uio_iov;
-  int iovcnt = uio->uio_iovcnt;
-  int i;
-
   UNUSED(filep);
 
-  for (i = 0; i < iovcnt; i++)
-    {
-      memset(iov[i].iov_base, 0, iov[i].iov_len);
-    }
-
-  uio_advance(uio, total);
-
-  return total;
+  memset(buffer, 0, len);
+  return len;
 }
 
 /****************************************************************************
- * Name: devzero_writev
+ * Name: devzero_write
  ****************************************************************************/
 
-static ssize_t devzero_writev(FAR struct file *filep, FAR struct uio *uio)
+static ssize_t devzero_write(FAR struct file *filep, FAR const char *buffer,
+                             size_t len)
 {
-  size_t total;
   UNUSED(filep);
+  UNUSED(buffer);
 
-  total = uio->uio_resid;
-
-  uio_advance(uio, total);
-  return total;
+  return len;
 }
 
 /****************************************************************************