## Summary

drivers/iovec: revert vector io implement from loop/null/zero driver

basic driver does not need such complex wrapper

Signed-off-by: chao an <anchao.archer@bytedance.com>


## Impact

improve the performance

## Testing

test code on ubuntu 24.04 x86_64

```
static void timespec_sub(struct timespec *dest,
                         struct timespec *ts1,
                         struct timespec *ts2)
{
  dest->tv_sec = ts1->tv_sec - ts2->tv_sec;
  dest->tv_nsec = ts1->tv_nsec - ts2->tv_nsec;

  if (dest->tv_nsec < 0)
    {
      dest->tv_nsec += 1000000000;
      dest->tv_sec -= 1;
    }
}


int main(int argc, FAR char *argv[])
{
  struct timespec result;
  struct timespec start;
  struct timespec end;
  int fd = open("/dev/zero", O_RDONLY);
  int loop = 0;
  int i = 0;
  char c;

  while (loop++ < 10)
    {
      i = 0;
      clock_gettime(CLOCK_MONOTONIC, &start);
      while (i++ < 100000)
        {
          read(fd, &c, 1);
        }
      clock_gettime(CLOCK_MONOTONIC, &end);

      timespec_sub(&result, &end, &start);
      printf("%d:spending %d.%lds\n", loop, result.tv_sec, result.tv_nsec);
    }

  close(fd);
  return 0;
}
```

before:
```
nsh> hello
1:spending 0.39044654s
2:spending 0.28245015s
3:spending 0.28215794s
4:spending 0.28436127s
5:spending 0.28330915s
6:spending 0.28564837s
7:spending 0.28116804s
8:spending 0.27876377s
9:spending 0.28266779s
10:spending 0.28328176s
nsh>
```



this PR + https://github.com/apache/nuttx/pull/15603
```
nsh> hello
1:spending 0.30323709s
2:spending 0.22735038s
3:spending 0.22735801s
4:spending 0.22715055s
5:spending 0.22756127s
6:spending 0.22938592s
7:spending 0.22860041s
8:spending 0.22718114s
9:spending 0.22785607s
10:spending 0.22676567s
nsh> poweroff
```